app/requirements.txt
streamlit==1.32.0
pandas==2.2.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openpyxl==3.1.2